@@ -1,6 +1,5 @@
-'use strict';
-
-const cloneDeep = require('rfdc')();
+import rfdc from 'rfdc';
+const cloneDeep = rfdc();
 
 class Document {
 
@@ -101,4 +100,4 @@ function isGetter(obj, key) {
   return Object.getOwnPropertyDescriptor(obj, key).get;
 }
 
-module.exports = Document;
+export default Document;
