@@ -1,5 +1,3 @@
-'use strict';
-
 class WarehouseError extends Error {
 
   /**
@@ -22,4 +20,4 @@ WarehouseError.ID_EXIST = 'ID_EXIST';
 WarehouseError.ID_NOT_EXIST = 'ID_NOT_EXIST';
 WarehouseError.ID_UNDEFINED = 'ID_UNDEFINED';
 
-module.exports = WarehouseError;
+export default WarehouseError;
