@@ -1,9 +1,7 @@
-'use strict';
-
-const WarehouseError = require('../error');
+import WarehouseError from '../error.js';
 
 class PopulationError extends WarehouseError {}
 
 PopulationError.prototype.name = 'PopulationError';
 
-module.exports = PopulationError;
+export default PopulationError;
