@@ -1,9 +1,7 @@
-'use strict';
-
-const WarehouseError = require('../error');
+import WarehouseError from '../error.js';
 
 class ValidationError extends WarehouseError {}
 
 ValidationError.prototype.name = 'ValidationError';
 
-module.exports = ValidationError;
+export default ValidationError;
