@@ -1,16 +1,15 @@
-'use strict';
-
-const { EventEmitter } = require('events');
-const cloneDeep = require('rfdc')();
-const Promise = require('bluebird');
-const { parseArgs, getProp, setGetter, shuffle } = require('./util');
-const Document = require('./document');
-const Query = require('./query');
-const Schema = require('./schema');
-const Types = require('./types');
-const WarehouseError = require('./error');
-const PopulationError = require('./error/population');
-const Mutex = require('./mutex');
+import { EventEmitter } from 'events';
+import rfdc from 'rfdc';
+const cloneDeep = rfdc();
+import Promise from 'bluebird';
+import { parseArgs, getProp, setGetter, shuffle } from './util.js';
+import Document from './document.js';
+import Query from './query.js';
+import Schema from './schema.js';
+import * as Types from './types/index.js';
+import WarehouseError from './error.js';
+import PopulationError from './error/population.js';
+import Mutex from './mutex.js';
 
 class Model extends EventEmitter {
 
@@ -963,4 +962,4 @@ Model.prototype.each = Model.prototype.forEach;
 
 Model.prototype.random = Model.prototype.shuffle;
 
-module.exports = Model;
+export default Model;
