@@ -1,5 +1,3 @@
-'use strict';
-
 class Mutex {
   constructor() {
     this._locked = false;
@@ -29,4 +27,4 @@ class Mutex {
   }
 }
 
-module.exports = Mutex;
+export default Mutex;
