@@ -1,7 +1,5 @@
-'use strict';
-
-const Promise = require('bluebird');
-const { parseArgs, shuffle } = require('./util');
+import Promise from 'bluebird';
+import { parseArgs, shuffle } from './util.js';
 
 class Query {
 
@@ -387,4 +385,4 @@ Query.prototype.each = Query.prototype.forEach;
 
 Query.prototype.random = Query.prototype.shuffle;
 
-module.exports = Query;
+export default Query;
