@@ -1,11 +1,9 @@
-'use strict';
-
-const SchemaType = require('./schematype');
-const Types = require('./types');
-const Promise = require('bluebird');
-const { getProp, setProp, delProp } = require('./util');
-const PopulationError = require('./error/population');
-const { isPlainObject } = require('is-plain-object');
+import SchemaType from './schematype.js';
+import * as Types from './types/index.js';
+import Promise from 'bluebird';
+import { getProp, setProp, delProp } from './util.js';
+import PopulationError from './error/population.js';
+import { isPlainObject } from 'is-plain-object';
 
 /**
  * @callback queryFilterCallback
@@ -792,4 +790,4 @@ class Schema {
 Schema.prototype.Types = Types;
 Schema.Types = Schema.prototype.Types;
 
-module.exports = Schema;
+export default Schema;
