@@ -1,7 +1,5 @@
-'use strict';
-
-const { setProp } = require('./util');
-const ValidationError = require('./error/validation');
+import { setProp } from './util.js';
+import ValidationError from './error/validation.js';
 
 /**
  * This is the basic schema type.
@@ -292,4 +290,4 @@ SchemaType.prototype.q$max = SchemaType.prototype.q$lte;
 
 SchemaType.prototype.q$min = SchemaType.prototype.q$gte;
 
-module.exports = SchemaType;
+export default SchemaType;
