@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 const { isArray } = Array;
 
@@ -378,4 +376,4 @@ SchemaTypeArray.prototype.u$append = SchemaTypeArray.prototype.u$push;
 
 SchemaTypeArray.prototype.u$prepend = SchemaTypeArray.prototype.u$unshift;
 
-module.exports = SchemaTypeArray;
+export default SchemaTypeArray;
