@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * Boolean schema type.
@@ -63,4 +61,4 @@ class SchemaTypeBoolean extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeBoolean;
+export default SchemaTypeBoolean;
