@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * Boolean schema type.
@@ -107,4 +105,4 @@ class SchemaTypeBuffer extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeBuffer;
+export default SchemaTypeBuffer;
