@@ -1,8 +1,6 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const cuid = require('cuid');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import cuid from 'cuid';
+import ValidationError from '../error/validation.js';
 
 /**
  * [CUID](https://github.com/ericelliott/cuid) schema type.
@@ -41,4 +39,4 @@ class SchemaTypeCUID extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeCUID;
+export default SchemaTypeCUID;
