@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * Date schema type.
@@ -154,4 +152,4 @@ class SchemaTypeDate extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeDate;
+export default SchemaTypeDate;
