@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * Enum schema type.
@@ -41,4 +39,4 @@ class SchemaTypeEnum extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeEnum;
+export default SchemaTypeEnum;
