@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaTypeNumber = require('./number');
-const ValidationError = require('../error/validation');
+import SchemaTypeNumber from './number.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * Integer schema type.
@@ -39,4 +37,4 @@ class SchemaTypeInteger extends SchemaTypeNumber {
   }
 }
 
-module.exports = SchemaTypeInteger;
+export default SchemaTypeInteger;
