@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * Number schema type.
@@ -125,4 +123,4 @@ class SchemaTypeNumber extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeNumber;
+export default SchemaTypeNumber;
