@@ -1,6 +1,4 @@
-'use strict';
-
-const SchemaType = require('../schematype');
+import SchemaType from '../schematype.js';
 
 /**
  * Object schema type.
@@ -19,4 +17,4 @@ class SchemaTypeObject extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeObject;
+export default SchemaTypeObject;
