@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const ValidationError = require('../error/validation');
+import SchemaType from '../schematype.js';
+import ValidationError from '../error/validation.js';
 
 /**
  * String schema type.
@@ -100,4 +98,4 @@ class SchemaTypeString extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeString;
+export default SchemaTypeString;
