@@ -1,7 +1,5 @@
-'use strict';
-
-const SchemaType = require('../schematype');
-const { setGetter } = require('../util');
+import SchemaType from '../schematype.js';
+import { setGetter } from '../util.js';
 
 /**
  * Virtual schema type.
@@ -77,4 +75,4 @@ class SchemaTypeVirtual extends SchemaType {
   }
 }
 
-module.exports = SchemaTypeVirtual;
+export default SchemaTypeVirtual;
