@@ -1,10 +1,9 @@
-'use strict';
-
-require('chai').should();
-const ValidationError = require('../../../lib/error/validation');
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import ValidationError from '../../../lib/error/validation.js';
+import SchemaTypeBoolean from '../../../lib/types/boolean.js';
 
 describe('SchemaTypeBoolean', () => {
-  const SchemaTypeBoolean = require('../../../lib/types/boolean');
   const type = new SchemaTypeBoolean('test');
 
   it('cast()', () => {
