@@ -1,10 +1,9 @@
-'use strict';
-
-const should = require('chai').should(); // eslint-disable-line
-const ValidationError = require('../../../lib/error/validation');
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import ValidationError from '../../../lib/error/validation.js';
+import SchemaTypeDate from '../../../lib/types/date.js';
 
 describe('SchemaTypeDate', () => {
-  const SchemaTypeDate = require('../../../lib/types/date');
   const type = new SchemaTypeDate('test');
 
   it('cast()', () => {
