@@ -1,10 +1,9 @@
-'use strict';
-
-require('chai').should();
-const ValidationError = require('../../../lib/error/validation');
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import ValidationError from '../../../lib/error/validation.js';
+import SchemaTypeInteger from '../../../lib/types/integer.js';
 
 describe('SchemaTypeInteger', () => {
-  const SchemaTypeInteger = require('../../../lib/types/integer');
   const type = new SchemaTypeInteger('test');
 
   it('cast()', () => {
