@@ -1,10 +1,9 @@
-'use strict';
-
-require('chai').should();
-const ValidationError = require('../../../lib/error/validation');
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import ValidationError from '../../../lib/error/validation.js';
+import SchemaTypeNumber from '../../../lib/types/number.js';
 
 describe('SchemaTypeNumber', () => {
-  const SchemaTypeNumber = require('../../../lib/types/number');
   const type = new SchemaTypeNumber('type');
 
   it('cast()', () => {
