@@ -1,16 +1,16 @@
-'use strict';
+import JSONStream from 'JSONStream';
+import Promise from 'bluebird';
+import fs from 'graceful-fs';
+import Model from './model.js';
+import Schema from './schema.js';
+import SchemaType from './schematype.js';
+import WarehouseError from './error.js';
+import pkg from '../package.json' assert { type: "json" };
+import log from 'hexo-log';
+import { Stream } from 'stream';
 
-const JSONStream = require('JSONStream');
-const Promise = require('bluebird');
-const fs = require('graceful-fs');
-const Model = require('./model');
-const Schema = require('./schema');
-const SchemaType = require('./schematype');
-const WarehouseError = require('./error');
-const pkg = require('../package.json');
+const pipeline = Promise.promisify(Stream.pipeline);
 const { open } = fs.promises;
-const pipeline = Promise.promisify(require('stream').pipeline);
-const log = require('hexo-log')();
 
 let _writev;
 
@@ -69,7 +69,7 @@ async function exportAsync(database, path) {
   }
 }
 
-class Database {
+export default class Database {
 
   /**
    * Database constructor.
@@ -189,5 +189,3 @@ Database.Schema = Database.prototype.Schema;
 Database.prototype.SchemaType = SchemaType;
 Database.SchemaType = Database.prototype.SchemaType;
 Database.version = pkg.version;
-
-module.exports = Database;
