@@ -1,14 +1,27 @@
-'use strict';
+import SchemaType from '../schematype.js';
+import SchemaTypeString from './string.js';
+import SchemaTypeNumber from './number.js';
+import SchemaTypeBoolean from './boolean.js';
+import SchemaTypeArray from './array.js';
+import SchemaTypeObject from './object.js';
+import SchemaTypeDate from './date.js';
+import SchemaTypeVirtual from './virtual.js';
+import SchemaTypeCUID from './cuid.js';
+import SchemaTypeEnum from './enum.js';
+import SchemaTypeInteger from './integer.js';
+import SchemaTypeBuffer from './buffer.js';
 
-exports.Mixed = require('../schematype');
-exports.String = require('./string');
-exports.Number = require('./number');
-exports.Boolean = require('./boolean');
-exports.Array = require('./array');
-exports.Object = require('./object');
-exports.Date = require('./date');
-exports.Virtual = require('./virtual');
-exports.CUID = require('./cuid');
-exports.Enum = require('./enum');
-exports.Integer = require('./integer');
-exports.Buffer = require('./buffer');
+export {
+  SchemaType as Mixed,
+  SchemaTypeString as String,
+  SchemaTypeNumber as Number,
+  SchemaTypeBoolean as Boolean,
+  SchemaTypeArray as Array,
+  SchemaTypeObject as Object,
+  SchemaTypeDate as Date,
+  SchemaTypeVirtual as Virtual,
+  SchemaTypeCUID as CUID,
+  SchemaTypeEnum as Enum,
+  SchemaTypeInteger as Integer,
+  SchemaTypeBuffer as Buffer
+};
