@@ -1,6 +1,4 @@
-'use strict';
-
-exports.shuffle = array => {
+export function shuffle(array) {
   if (!Array.isArray(array)) throw new TypeError('array must be an Array!');
   const $array = array.slice();
   const { length } = $array;
@@ -17,13 +15,13 @@ exports.shuffle = array => {
   }
 
   return $array;
-};
+}
 
 function extractPropKey(key) {
   return key.split('.');
 }
 
-exports.getProp = (obj, key) => {
+export function getProp(obj, key) {
   if (typeof obj !== 'object') throw new TypeError('obj must be an object!');
   if (!key) throw new TypeError('key is required!');
 
@@ -40,9 +38,9 @@ exports.getProp = (obj, key) => {
   }
 
   return result;
-};
+}
 
-exports.setProp = (obj, key, value) => {
+export function setProp(obj, key, value) {
   if (typeof obj !== 'object') throw new TypeError('obj must be an object!');
   if (!key) throw new TypeError('key is required!');
 
@@ -64,9 +62,9 @@ exports.setProp = (obj, key, value) => {
   }
 
   cursor[lastKey] = value;
-};
+}
 
-exports.delProp = (obj, key) => {
+export function delProp(obj, key) {
   if (typeof obj !== 'object') throw new TypeError('obj must be an object!');
   if (!key) throw new TypeError('key is required!');
 
@@ -92,9 +90,9 @@ exports.delProp = (obj, key) => {
   }
 
   delete cursor[lastKey];
-};
+}
 
-exports.setGetter = (obj, key, fn) => {
+export function setGetter(obj, key, fn) {
   if (typeof obj !== 'object') throw new TypeError('obj must be an object!');
   if (!key) throw new TypeError('key is required!');
   if (typeof fn !== 'function') throw new TypeError('fn must be a function!');
@@ -117,9 +115,9 @@ exports.setGetter = (obj, key, fn) => {
   }
 
   cursor.__defineGetter__(lastKey, fn);
-};
+}
 
-exports.arr2obj = (arr, value) => {
+export function arr2obj(arr, value) {
   if (!Array.isArray(arr)) throw new TypeError('arr must be an array!');
 
   const obj = {};
@@ -130,9 +128,9 @@ exports.arr2obj = (arr, value) => {
   }
 
   return obj;
-};
+}
 
-exports.reverse = arr => {
+export function reverse(arr) {
   if (!Array.isArray(arr)) throw new TypeError('arr must be an array!');
 
   const len = arr.length;
@@ -146,9 +144,9 @@ exports.reverse = arr => {
   }
 
   return arr;
-};
+}
 
-function parseArgs(args) {
+function _parseArgs(args) {
   if (typeof args !== 'string') return args;
 
   const arr = args.split(' ');
@@ -174,17 +172,17 @@ function parseArgs(args) {
   return result;
 }
 
-exports.parseArgs = (orderby, order) => {
+export function parseArgs(orderby, order) {
   let result;
 
   if (order) {
     result = {};
     result[orderby] = order;
   } else if (typeof orderby === 'string') {
-    result = parseArgs(orderby);
+    result = _parseArgs(orderby);
   } else {
     result = orderby;
   }
 
   return result;
-};
+}
