@@ -1,9 +1,8 @@
-'use strict';
-
-require('chai').should();
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import Database from '../../lib/database.js';
 
 describe('Schema', () => {
-  const Database = require('../..');
   const Schema = Database.Schema;
 
   it('add()', () => {
